requirements.txt
appdirs==1.4.4
Appium-Python-Client==1.1.0
asgiref
backcall==0.2.0
beautifulsoup4==4.10.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
configparser==5.0.2
contextlib2==0.6.0
crayons==0.4.0
decorator==5.0.9
distlib==0.3.0
distro==1.4.0
Django==3.2.3
django-admin-rangefilter==0.8.4
django-extensions==3.1.3
et-xmlfile==1.1.0
Faker
filelock==3.0.12
html5lib==1.0.1
huggingface-hub==0.0.17
HyperPyYAML==1.0.0
idna==2.8
importlib-metadata==4.0.1
ipaddr==2.2.0
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
lockfile==0.12.2
matplotlib-inline==0.1.2
msgpack==0.6.2
mypy-extensions==0.4.3
numpy==1.21.2
openpyxl==3.0.9
packaging==21.0
parso==0.8.2
pathspec==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
progress==1.5
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-python-adb==0.3.0.dev0
pydub==0.25.1
Pygments==2.9.0
pyparsing==2.4.6
python-dateutil==2.8.2
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
retrying==1.3.3
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scipy==1.7.1
selenium==3.141.0
sentencepiece==0.1.96
six==1.14.0
soupsieve==2.2.1
speechbrain==0.5.10
SpeechRecognition==3.8.1
sqlparse==0.4.1
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
tqdm==4.62.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.5
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
zipp==3.4.1
